Updated the methods.ts to be a directory w/ functions in separate files.
